Rendering of images implemented. Connected to FileIO load and save.
This build now depends on a pull request I've made to LibSndFile so that both images and sounds can be easily loaded.

Also, there is some commented out code that starts to try to connect this package to Compose.jl. I'm running into problems initializing Cairo. When these two are included, opening an SDL window leads to a Mac OS GUI exception.