Updated linux error message, change timeline for linux support.
After investigating how easy it would be to get linux working for
Weber it doesn't appear to be very straightforward. I am getting
strange LLVM errors, and the standard method for installing
dependencies on linux is currently broken. At the very least I need to
wait for the Julia dev team to fix this problem before I can
proceede. I would then need to track down the source of the LLVM
error, and that is likely to invovle yet another bug in julia itself,
I am guessing.