Improved test of compound moment timing.
As written before, the test could fail due to GC occuring at the end
of a trial, but before the final part of a compound moment was
completed. This is an edge case we really don't care about... it only
occurs when a trial moment is longer than the trial length itself. It
does suggest that maybe one feature of Weber should be to end a trial
when all moments are complete??? But maybe someone would want this for
some reason (e.g. my "Faster!" message).