Call prepare! just before the first non-zero pause before a moment
This allows a string of delta_t zero moments to occur between prepare! and the
moment prepared, thus increasing the number of cases where prepare! is called
some short time before the moment it preapres is run (rather than immedaitely
before that moment runs).

This functionality is used to allow sounds to be prepared ahead of time, and
thus precisely timed, in some common uses cases that weren't covered by the
previosu functionality.

Closes #23